refactor: handle signup errors
