+++ frontend/src/api/auth/constants.ts
@@ -1,0 +1,3 @@
+export enum SIGNUP_ERROR {
+  USERNAME_ALREADY_EXISTS = 'USERNAME_ALREADY_EXISTS',
+}
